Log元件更新

+++ Client/Assets/Plugins/GoogleAnalyticsV3/GoogleAnalyticsAndroidV3.cs
@@ -31,7 +31,7 @@ public class GoogleAnalyticsAndroidV3 : IDisposable {
   private string appName;
   private string bundleIdentifier;
   private int dispatchPeriod;
-  private int sampleFrequency;
+  private float sampleFrequency;
   private GoogleAnalyticsV3.DebugMode logLevel;
   private bool anonymizeIP;
   private bool dryRun;
@@ -89,7 +89,7 @@ public class GoogleAnalyticsAndroidV3 : IDisposable {
     tracker.Call(GoogleAnalyticsV3.SET, args);
   }
 
-  public void SetSampleFrequency(int sampleFrequency) {
+  public void SetSampleFrequency(float sampleFrequency) {
     this.sampleFrequency = sampleFrequency;
   }
 
@@ -350,6 +350,7 @@ public class GoogleAnalyticsAndroidV3 : IDisposable {
     object[] args;
     if (builder.GetCurrencyCode() != null) {
       args = new object[7];
+      // TODO: Validate currency code
       args[6] = builder.GetCurrencyCode();
     } else {
       args = new object[6];
